remove PPW endpoints and code
