+++ core/src/main/scala/org/broadinstitute/dsde/rawls/workspace/WorkspaceService.scala
@@ -38,7 +38,6 @@ import org.broadinstitute.dsde.rawls.model.WorkspaceType.WorkspaceType
 import org.broadinstitute.dsde.rawls.model.WorkspaceVersions.WorkspaceVersion
 import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.monitor.migration.MigrationUtils.Implicits.monadThrowDBIOAction
-import org.broadinstitute.dsde.rawls.monitor.migration.WorkspaceMigrationMetadata
 import org.broadinstitute.dsde.rawls.resourcebuffer.ResourceBufferService
 import org.broadinstitute.dsde.rawls.serviceperimeter.ServicePerimeterService
 import org.broadinstitute.dsde.rawls.user.UserService
@@ -3137,50 +3136,6 @@ class WorkspaceService(protected val ctx: RawlsRequestContext,
       servicePerimeterService.overwriteGoogleProjectsInPerimeter(servicePerimeterName, dataAccess)
     }
 
-  // Admin endpoint, not limited to V2 workspaces
-  def getWorkspaceMigrationAttempts(workspaceName: WorkspaceName): Future[List[WorkspaceMigrationMetadata]] =
-    asFCAdmin {
-      for {
-        workspace <- getWorkspaceContext(workspaceName)
-        attempts <- dataSource.inTransaction { dataAccess =>
-          dataAccess.workspaceMigrationQuery.getMigrationAttempts(workspace)
-        }
-      } yield attempts.mapWithIndex(WorkspaceMigrationMetadata.fromWorkspaceMigration)
-    }
-
-  // Admin endpoint, not limited to V2 workspaces
-  def migrateWorkspace(workspaceName: WorkspaceName): Future[WorkspaceMigrationMetadata] =
-    asFCAdmin {
-      logger.info(s"Scheduling Workspace '$workspaceName' for migration")
-      dataSource.inTransaction { dataAccess =>
-        dataAccess.workspaceMigrationQuery.scheduleAndGetMetadata(workspaceName)
-      }
-    }
-
-  // Admin endpoint, not limited to V2 workspaces
-  def migrateAll(workspaceNames: Iterable[WorkspaceName]): Future[Iterable[WorkspaceMigrationMetadata]] =
-    asFCAdmin {
-      dataSource.inTransaction { dataAccess =>
-        for {
-          errorsOrMigrationAttempts <- workspaceNames.toList.traverse { workspaceName =>
-            MonadThrow[ReadWriteAction].attempt {
-              dataAccess.workspaceMigrationQuery.scheduleAndGetMetadata(workspaceName)
-            }
-          }
-
-          (errors, migrationAttempts) = errorsOrMigrationAttempts.partitionMap(identity)
-          _ <- MonadThrow[ReadWriteAction].raiseUnless(errors.isEmpty) {
-            new RawlsExceptionWithErrorReport(
-              ErrorReport(StatusCodes.BadRequest,
-                          "One or more workspaces could not be scheduled for migration",
-                          errors.map(ErrorReport.apply)
-              )
-            )
-          }
-
-        } yield migrationAttempts
-      }
-    }
 
   private def failUnlessBillingAccountHasAccess(billingProject: RawlsBillingProject,
                                                 parentContext: RawlsRequestContext = ctx
