+++ core/src/test/scala/org/broadinstitute/dsde/rawls/workspace/WorkspaceServiceSpec.scala
@@ -7,7 +7,7 @@ import akka.http.scaladsl.testkit.ScalatestRouteTest
 import bio.terra.profile.model.ProfileModel
 import bio.terra.workspace.client.ApiException
 import bio.terra.workspace.model.{AzureContext, GcpContext, WorkspaceDescription, WsmPolicyInput, WsmPolicyPair}
-import cats.implicits.{catsSyntaxOptionId, toTraverseOps}
+import cats.implicits.{catsSyntaxOptionId}
 import com.google.api.client.googleapis.json.{GoogleJsonError, GoogleJsonResponseException}
 import com.google.api.client.http.{HttpHeaders, HttpResponseException}
 import com.google.api.services.cloudresourcemanager.model.Project
@@ -34,7 +34,6 @@ import org.broadinstitute.dsde.rawls.model.AttributeUpdateOperations._
 import org.broadinstitute.dsde.rawls.model.ProjectPoolType.ProjectPoolType
 import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport._
 import org.broadinstitute.dsde.rawls.model._
-import org.broadinstitute.dsde.rawls.monitor.migration.MigrationUtils.Implicits.monadThrowDBIOAction
 import org.broadinstitute.dsde.rawls.openam.MockUserInfoDirectivesWithUser
 import org.broadinstitute.dsde.rawls.resourcebuffer.ResourceBufferService
 import org.broadinstitute.dsde.rawls.serviceperimeter.ServicePerimeterService
@@ -43,7 +42,6 @@ import org.broadinstitute.dsde.rawls.util.MockitoTestUtils
 import org.broadinstitute.dsde.rawls.webservice._
 import org.broadinstitute.dsde.rawls.{
   NoSuchWorkspaceException,
-  RawlsException,
   RawlsExceptionWithErrorReport,
   RawlsTestUtils
 }
@@ -2026,64 +2024,6 @@ class WorkspaceServiceSpec
       }
     }
   }
-
-  "migrateWorkspace" should "create an entry in the migration table" in
-    withTestDataServices { services =>
-      Await.result(
-        for {
-          _ <- services.workspaceService.migrateWorkspace(testData.v1Workspace.toWorkspaceName)
-          isMigrating <- services.slickDataSource.inTransaction { dataAccess =>
-            dataAccess.workspaceMigrationQuery.isPendingMigration(testData.v1Workspace)
-          }
-        } yield isMigrating should be(true),
-        30.seconds
-      )
-    }
-
-  "migrateAll" should "create and entry for each workspace listed" in
-    withTestDataServices { services =>
-      val workspaces = List(testData.v1Workspace)
-      Await.result(
-        for {
-          _ <- services.workspaceService.migrateAll(workspaces.map(_.toWorkspaceName))
-          isPendingMigration <- services.slickDataSource.inTransaction { dataAccess =>
-            workspaces.traverse(dataAccess.workspaceMigrationQuery.isPendingMigration)
-          }
-        } yield every(isPendingMigration) shouldBe true,
-        30.seconds
-      )
-    }
-
-  it should "not schedule any migrations if workspace is invalid" in
-    withTestDataServices { services =>
-      val workspaces = List(testData.v1Workspace, testData.workspace)
-      intercept[RawlsExceptionWithErrorReport] {
-        Await.result(services.workspaceService.migrateAll(workspaces.map(_.toWorkspaceName)), 30.seconds)
-      }
-
-      val isPendingMigration = Await.result(services.slickDataSource.inTransaction { dataAccess =>
-                                              workspaces.traverse(dataAccess.workspaceMigrationQuery.isPendingMigration)
-                                            },
-                                            30.seconds
-      )
-
-      every(isPendingMigration) shouldBe false
-    }
-
-  "getWorkspaceMigrations" should "return a list of workspace migration attempts" in
-    withTestDataServices { services =>
-      Await.result(
-        for {
-          before <- services.workspaceService.getWorkspaceMigrationAttempts(testData.v1Workspace.toWorkspaceName)
-          _ <- services.workspaceService.migrateWorkspace(testData.v1Workspace.toWorkspaceName)
-          after <- services.workspaceService.getWorkspaceMigrationAttempts(testData.v1Workspace.toWorkspaceName)
-        } yield {
-          before shouldBe empty
-          after should not be empty
-        },
-        30.seconds
-      )
-    }
 
   "createWorkspace" should "create a V2 Workspace" in withTestDataServices { services =>
     val newWorkspaceName = "space_for_workin"
